Move admin user-y endpoints to /api/admin
