+++ src/main/resources/swagger/rawls.yaml
@@ -606,6 +606,122 @@ paths:
             - openid
             - email
             - profile
+  '/api/admin/user/{userSubjectId}/enable':
+       post:
+         responses:
+           '204':
+             description: Successful Request
+           '403':
+             description: Only administrators can enable users
+           '500':
+             description: Rawls Internal Error
+             schema:
+               $ref: '#/definitions/ErrorReport'
+         description: "creates a user in the Rawls system"
+         tags:
+           - admin
+           - users
+         summary: Enable user
+         operationId: enable
+         parameters:
+          - in: path
+            description: subject id of user
+            name: userSubjectId
+            required: true
+            type: string
+         security:
+           - authorization:
+              - openid
+              - email
+              - profile
+  '/api/admin/user/{userSubjectId}/disable':
+     post:
+       responses:
+         '204':
+           description: Successful Request
+         '403':
+           description: Only administrators can disable users
+         '500':
+           description: Rawls Internal Error
+           schema:
+             $ref: '#/definitions/ErrorReport'
+       description: "disables a user in the Rawls system"
+       tags:
+         - admin
+         - users
+       summary: Disable user
+       operationId: disable
+       parameters:
+        - in: path
+          description: subject id of user
+          name: userSubjectId
+          required: true
+          type: string
+       security:
+         - authorization:
+            - openid
+            - email
+            - profile
+  '/api/admin/user/{userSubjectId}':
+       delete:
+         responses:
+           '204':
+             description: Successful Request
+           '400':
+             description: Cannot delete user with submissions
+           '403':
+             description: Only administrators can delete users
+           '500':
+             description: Rawls Internal Error
+             schema:
+               $ref: '#/definitions/ErrorReport'
+         description: "Deletes a user in the Rawls system.  This endpoint should only be used with test users who will never have submissions."
+         tags:
+           - admin
+           - users
+         summary: Delete user
+         operationId: delete
+         parameters:
+          - in: path
+            description: subject id of user
+            name: userSubjectId
+            required: true
+            type: string
+         security:
+           - authorization:
+              - openid
+              - email
+              - profile
+  '/api/admin/user/{userSubjectId}':
+     get:
+       responses:
+         '200':
+           description: Successful Request
+           schema:
+             $ref: '#/definitions/UserStatus'
+         '403':
+           description: You must be an administrator to get user status
+         '500':
+           description: Rawls Internal Error
+           schema:
+             $ref: '#/definitions/ErrorReport'
+       description: "gets the status of a user"
+       tags:
+         - admin
+         - users
+       summary: Get status of user
+       operationId: adminStatus
+       parameters:
+        - in: path
+          description: subject id of user
+          name: userSubjectId
+          required: true
+          type: string
+       security:
+         - authorization:
+            - openid
+            - email
+            - profile
   '/api/admin/allUserReadAccess/{workspaceNamespace}/{workspaceName}':
     get:
       responses:
@@ -2693,92 +2809,6 @@ paths:
             - openid
             - email
             - profile
-  '/api/user/{userSubjectId}/enable':
-       post:
-         responses:
-           '204':
-             description: Successful Request
-           '403':
-             description: Only administrators can enable users
-           '500':
-             description: Rawls Internal Error
-             schema:
-               $ref: '#/definitions/ErrorReport'
-         description: "creates a user in the Rawls system"
-         tags:
-           - admin
-           - users
-         summary: Enable user
-         operationId: enable
-         parameters:
-          - in: path
-            description: subject id of user
-            name: userSubjectId
-            required: true
-            type: string
-         security:
-           - authorization:
-              - openid
-              - email
-              - profile
-  '/api/user/{userSubjectId}/disable':
-     post:
-       responses:
-         '204':
-           description: Successful Request
-         '403':
-           description: Only administrators can disable users
-         '500':
-           description: Rawls Internal Error
-           schema:
-             $ref: '#/definitions/ErrorReport'
-       description: "disables a user in the Rawls system"
-       tags:
-         - admin
-         - users
-       summary: Disable user
-       operationId: disable
-       parameters:
-        - in: path
-          description: subject id of user
-          name: userSubjectId
-          required: true
-          type: string
-       security:
-         - authorization:
-            - openid
-            - email
-            - profile
-  '/api/user/{userSubjectId}':
-       delete:
-         responses:
-           '204':
-             description: Successful Request
-           '400':
-             description: Cannot delete user with submissions
-           '403':
-             description: Only administrators can delete users
-           '500':
-             description: Rawls Internal Error
-             schema:
-               $ref: '#/definitions/ErrorReport'
-         description: "Deletes a user in the Rawls system.  This endpoint should only be used with test users who will never have submissions."
-         tags:
-           - admin
-           - users
-         summary: Delete user
-         operationId: delete
-         parameters:
-          - in: path
-            description: subject id of user
-            name: userSubjectId
-            required: true
-            type: string
-         security:
-           - authorization:
-              - openid
-              - email
-              - profile
     '/api/user':
        get:
          responses:
@@ -2800,36 +2830,6 @@ paths:
               - openid
               - email
               - profile
-  '/api/user/{userSubjectId}':
-     get:
-       responses:
-         '200':
-           description: Successful Request
-           schema:
-             $ref: '#/definitions/UserStatus'
-         '403':
-           description: You must be an administrator to get user status
-         '500':
-           description: Rawls Internal Error
-           schema:
-             $ref: '#/definitions/ErrorReport'
-       description: "gets the status of a user"
-       tags:
-         - admin
-         - users
-       summary: Get status of user
-       operationId: adminStatus
-       parameters:
-        - in: path
-          description: subject id of user
-          name: userSubjectId
-          required: true
-          type: string
-       security:
-         - authorization:
-            - openid
-            - email
-            - profile
   '/api/user/refreshToken':
     put:
       responses:
